Changed store structure
