+++ src/components/Header/index.tsx
@@ -1,7 +1,7 @@
 import Button from '@/components/Button';
 import { Link } from 'react-router-dom';
 import { useStoreDispatch } from '@/store/store';
-import { toClearState } from '@/store/cardsSlice';
+import { toClearState } from '@/store/cards';
 import styles from '@/components/Header/header.module.scss';
 import { ROUTES } from '@/store/routes';
 import { FC } from 'react';
