+++ src/components/modal/Container/index.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import styles from '@/components/modal/Container/container.module.scss';
-import { ContainerProps } from '@/store/interfaces';
+import { ContainerProps } from '@/store/cards/cards.types';
 import { useSelector } from 'react-redux';
 import { RootState } from '@/store/store';
