+++ src/store/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { useDispatch } from 'react-redux';
-import cardsReducer from '@/store/cardsSlice';
-import { cardsAPI } from './apiSlice';
+import cardsReducer from '@/store/cards';
+import { cardsAPI } from './cards/cards.apiCalls';
 
 export const store = configureStore({
   reducer: {
